Split up Platform.c into Platform_WinApi.c, Platform_Web.c, and Platform_Posix.c
