+++ src/Platform_Web.c
@@ -1,0 +1,409 @@
+#include "Core.h"
+#if defined CC_BUILD_WEB
+
+#include "_PlatformBase.h"
+#include "Stream.h"
+#include "ExtMath.h"
+#include "Drawer2D.h"
+#include "Funcs.h"
+#include "Window.h"
+#include "Utils.h"
+#include "Errors.h"
+
+/* POSIX can be shared between Linux/BSD/macOS */
+#include <errno.h>
+#include <time.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+#include <dirent.h>
+#include <fcntl.h>
+#include <arpa/inet.h>
+#include <netinet/in.h>
+#include <sys/socket.h>
+#include <sys/ioctl.h>
+#include <sys/types.h>
+#include <sys/stat.h>
+#include <sys/time.h>
+#include <utime.h>
+#include <signal.h>
+#include <stdio.h>
+
+const cc_result ReturnCode_FileShareViolation = 1000000000; /* TODO: not used apparently */
+const cc_result ReturnCode_FileNotFound     = ENOENT;
+const cc_result ReturnCode_SocketInProgess  = EINPROGRESS;
+const cc_result ReturnCode_SocketWouldBlock = EWOULDBLOCK;
+const cc_result ReturnCode_DirectoryExists  = EEXIST;
+#include <emscripten.h>
+#include "Chat.h"
+
+
+/*########################################################################################################################*
+*---------------------------------------------------------Memory----------------------------------------------------------*
+*#########################################################################################################################*/
+void Mem_Set(void*  dst, cc_uint8 value,  cc_uint32 numBytes) { memset(dst, value, numBytes); }
+void Mem_Copy(void* dst, const void* src, cc_uint32 numBytes) { memcpy(dst, src,   numBytes); }
+
+void* Mem_TryAlloc(cc_uint32 numElems, cc_uint32 elemsSize) {
+	cc_uint32 size = CalcMemSize(numElems, elemsSize);
+	return size ? malloc(size) : NULL;
+}
+
+void* Mem_TryAllocCleared(cc_uint32 numElems, cc_uint32 elemsSize) {
+	return calloc(numElems, elemsSize);
+}
+
+void* Mem_TryRealloc(void* mem, cc_uint32 numElems, cc_uint32 elemsSize) {
+	cc_uint32 size = CalcMemSize(numElems, elemsSize);
+	return size ? realloc(mem, size) : NULL;
+}
+
+void Mem_Free(void* mem) {
+	if (mem) free(mem);
+}
+
+
+/*########################################################################################################################*
+*------------------------------------------------------Logging/Time-------------------------------------------------------*
+*#########################################################################################################################*/
+/* TODO: check this is actually accurate */
+static cc_uint64 sw_freqMul = 1, sw_freqDiv = 1;
+cc_uint64 Stopwatch_ElapsedMicroseconds(cc_uint64 beg, cc_uint64 end) {
+	if (end < beg) return 0;
+	return ((end - beg) * sw_freqMul) / sw_freqDiv;
+}
+
+void Platform_Log(const char* msg, int len) {
+	write(STDOUT_FILENO, msg,  len);
+	write(STDOUT_FILENO, "\n",   1);
+}
+
+#define UnixTime_TotalMS(time) ((cc_uint64)time.tv_sec * 1000 + UNIX_EPOCH + (time.tv_usec / 1000))
+TimeMS DateTime_CurrentUTC_MS(void) {
+	struct timeval cur;
+	gettimeofday(&cur, NULL);
+	return UnixTime_TotalMS(cur);
+}
+
+void DateTime_CurrentLocal(struct DateTime* t) {
+	struct timeval cur; 
+	struct tm loc_time;
+	gettimeofday(&cur, NULL);
+	localtime_r(&cur.tv_sec, &loc_time);
+
+	t->year   = loc_time.tm_year + 1900;
+	t->month  = loc_time.tm_mon  + 1;
+	t->day    = loc_time.tm_mday;
+	t->hour   = loc_time.tm_hour;
+	t->minute = loc_time.tm_min;
+	t->second = loc_time.tm_sec;
+}
+
+cc_uint64 Stopwatch_Measure(void) {
+	/* time is a milliseconds double */
+	return (cc_uint64)(emscripten_get_now() * 1000);
+}
+
+
+/*########################################################################################################################*
+*-----------------------------------------------------Directory/File------------------------------------------------------*
+*#########################################################################################################################*/
+cc_result Directory_Create(const cc_string* path) {
+	char str[NATIVE_STR_LEN];
+	Platform_EncodeUtf8(str, path);
+	/* read/write/search permissions for owner and group, and with read/search permissions for others. */
+	/* TODO: Is the default mode in all cases */
+	return mkdir(str, S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH) == -1 ? errno : 0;
+}
+
+int File_Exists(const cc_string* path) {
+	char str[NATIVE_STR_LEN];
+	struct stat sb;
+	Platform_EncodeUtf8(str, path);
+	return stat(str, &sb) == 0 && S_ISREG(sb.st_mode);
+}
+
+cc_result Directory_Enum(const cc_string* dirPath, void* obj, Directory_EnumCallback callback) {
+	cc_string path; char pathBuffer[FILENAME_SIZE];
+	char str[NATIVE_STR_LEN];
+	DIR* dirPtr;
+	struct dirent* entry;
+	char* src;
+	int len, res;
+
+	Platform_EncodeUtf8(str, dirPath);
+	dirPtr = opendir(str);
+	if (!dirPtr) return errno;
+
+	/* POSIX docs: "When the end of the directory is encountered, a null pointer is returned and errno is not changed." */
+	/* errno is sometimes leftover from previous calls, so always reset it before readdir gets called */
+	errno = 0;
+	String_InitArray(path, pathBuffer);
+
+	while ((entry = readdir(dirPtr))) {
+		path.length = 0;
+		String_Format1(&path, "%s/", dirPath);
+
+		/* ignore . and .. entry */
+		src = entry->d_name;
+		if (src[0] == '.' && src[1] == '\0') continue;
+		if (src[0] == '.' && src[1] == '.' && src[2] == '\0') continue;
+
+		len = String_Length(src);
+		String_AppendUtf8(&path, src, len);
+
+		/* TODO: fallback to stat when this fails */
+		if (entry->d_type == DT_DIR) {
+			res = Directory_Enum(&path, obj, callback);
+			if (res) { closedir(dirPtr); return res; }
+		} else {
+			callback(&path, obj);
+		}
+		errno = 0;
+	}
+
+	res = errno; /* return code from readdir */
+	closedir(dirPtr);
+	return res;
+}
+
+static cc_result File_Do(cc_file* file, const cc_string* path, int mode) {
+	char str[NATIVE_STR_LEN];
+	Platform_EncodeUtf8(str, path);
+	*file = open(str, mode, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
+	return *file == -1 ? errno : 0;
+}
+
+cc_result File_Open(cc_file* file, const cc_string* path) {
+	return File_Do(file, path, O_RDONLY);
+}
+cc_result File_Create(cc_file* file, const cc_string* path) {
+	return File_Do(file, path, O_RDWR | O_CREAT | O_TRUNC);
+}
+cc_result File_OpenOrCreate(cc_file* file, const cc_string* path) {
+	return File_Do(file, path, O_RDWR | O_CREAT);
+}
+
+cc_result File_Read(cc_file file, void* data, cc_uint32 count, cc_uint32* bytesRead) {
+	*bytesRead = read(file, data, count);
+	return *bytesRead == -1 ? errno : 0;
+}
+
+cc_result File_Write(cc_file file, const void* data, cc_uint32 count, cc_uint32* bytesWrote) {
+	*bytesWrote = write(file, data, count);
+	return *bytesWrote == -1 ? errno : 0;
+}
+
+extern void interop_SyncFS(void);
+cc_result File_Close(cc_file file) {	
+	int res = close(file) == -1 ? errno : 0; 
+	interop_SyncFS(); 
+	return res;
+}
+
+cc_result File_Seek(cc_file file, int offset, int seekType) {
+	static cc_uint8 modes[3] = { SEEK_SET, SEEK_CUR, SEEK_END };
+	return lseek(file, offset, modes[seekType]) == -1 ? errno : 0;
+}
+
+cc_result File_Position(cc_file file, cc_uint32* pos) {
+	*pos = lseek(file, 0, SEEK_CUR);
+	return *pos == -1 ? errno : 0;
+}
+
+cc_result File_Length(cc_file file, cc_uint32* len) {
+	struct stat st;
+	if (fstat(file, &st) == -1) { *len = -1; return errno; }
+	*len = st.st_size; return 0;
+}
+
+
+/*########################################################################################################################*
+*--------------------------------------------------------Threading--------------------------------------------------------*
+*#########################################################################################################################*/
+/* No real threading support with emscripten backend */
+void Thread_Sleep(cc_uint32 milliseconds) { }
+void* Thread_Start(Thread_StartFunc func) { func(); return NULL; }
+void Thread_Detach(void* handle) { }
+void Thread_Join(void* handle) { }
+
+void* Mutex_Create(void) { return NULL; }
+void Mutex_Free(void* handle) { }
+void Mutex_Lock(void* handle) { }
+void Mutex_Unlock(void* handle) { }
+
+void* Waitable_Create(void) { return NULL; }
+void Waitable_Free(void* handle) { }
+void Waitable_Signal(void* handle) { }
+void Waitable_Wait(void* handle) { }
+void Waitable_WaitFor(void* handle, cc_uint32 milliseconds) { }
+
+
+/*########################################################################################################################*
+*--------------------------------------------------------Font/Text--------------------------------------------------------*
+*#########################################################################################################################*/
+void Platform_LoadSysFonts(void) { }
+
+
+/*########################################################################################################################*
+*---------------------------------------------------------Socket----------------------------------------------------------*
+*#########################################################################################################################*/
+cc_result Socket_Create(cc_socket* s) {
+	*s = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
+	return *s == -1 ? errno : 0;
+}
+
+cc_result Socket_Available(cc_socket s, int* available) {
+	return ioctl(s, FIONREAD, available);
+}
+cc_result Socket_SetBlocking(cc_socket s, cc_bool blocking) {
+	return ERR_NOT_SUPPORTED; /* sockets always async */
+}
+
+cc_result Socket_GetError(cc_socket s, cc_result* result) {
+	socklen_t resultSize = sizeof(cc_result);
+	return getsockopt(s, SOL_SOCKET, SO_ERROR, result, &resultSize);
+}
+
+cc_result Socket_Connect(cc_socket s, const cc_string* ip, int port) {
+	struct sockaddr addr;
+	cc_result res;
+	addr.sa_family = AF_INET;
+
+	Stream_SetU16_BE( (cc_uint8*)&addr.sa_data[0], port);
+	if (!Utils_ParseIP(ip, (cc_uint8*)&addr.sa_data[2])) 
+		return ERR_INVALID_ARGUMENT;
+
+	res = connect(s, &addr, sizeof(addr));
+	return res == -1 ? errno : 0;
+}
+
+cc_result Socket_Read(cc_socket s, cc_uint8* data, cc_uint32 count, cc_uint32* modified) {
+	/* recv only reads one WebSocket frame at most, hence call it multiple times */
+	int recvCount = 0, pending;
+	*modified = 0;
+
+	while (count && !Socket_Available(s, &pending) && pending) {
+		recvCount = recv(s, data, count, 0);
+		if (recvCount == -1) return errno;
+
+		*modified += recvCount;
+		data      += recvCount; count -= recvCount;
+	}
+	return 0;
+}
+
+cc_result Socket_Write(cc_socket s, const cc_uint8* data, cc_uint32 count, cc_uint32* modified) {
+	int sentCount = send(s, data, count, 0);
+	if (sentCount != -1) { *modified = sentCount; return 0; }
+	*modified = 0; return errno;
+}
+
+cc_result Socket_Close(cc_socket s) {
+	cc_result res = close(s);
+	if (res == -1) res = errno;
+	return res;
+}
+
+#include <poll.h>
+cc_result Socket_Poll(cc_socket s, int mode, cc_bool* success) {
+	struct pollfd pfd;
+	int flags;
+
+	pfd.fd     = s;
+	pfd.events = mode == SOCKET_POLL_READ ? POLLIN : POLLOUT;
+	if (poll(&pfd, 1, 0) == -1) { *success = false; return errno; }
+	
+	/* to match select, closed socket still counts as readable */
+	flags    = mode == SOCKET_POLL_READ ? (POLLIN | POLLHUP) : POLLOUT;
+	*success = (pfd.revents & flags) != 0;
+	return 0;
+}
+
+
+/*########################################################################################################################*
+*-----------------------------------------------------Process/Module------------------------------------------------------*
+*#########################################################################################################################*/
+cc_result Process_StartGame(const cc_string* args) { return ERR_NOT_SUPPORTED; }
+void Process_Exit(cc_result code) { exit(code); }
+
+extern int interop_OpenTab(const char* url);
+cc_result Process_StartOpen(const cc_string* args) {
+	char str[NATIVE_STR_LEN];
+	Platform_EncodeUtf8(str, args);
+	return interop_OpenTab(str);
+}
+
+
+/*########################################################################################################################*
+*--------------------------------------------------------Updater----------------------------------------------------------*
+*#########################################################################################################################*/
+const char* const Updater_D3D9 = NULL;
+const char* const Updater_OGL  = NULL;
+
+cc_result Updater_GetBuildTime(cc_uint64* t)   { return ERR_NOT_SUPPORTED; }
+cc_bool Updater_Clean(void)                    { return true; }
+cc_result Updater_Start(const char** action)   { return ERR_NOT_SUPPORTED; }
+cc_result Updater_MarkExecutable(void)         { return 0; }
+cc_result Updater_SetNewBuildTime(cc_uint64 t) { return ERR_NOT_SUPPORTED; }
+
+
+/*########################################################################################################################*
+*-------------------------------------------------------Dynamic lib-------------------------------------------------------*
+*#########################################################################################################################*/
+void* DynamicLib_Load2(const cc_string* path)      { return NULL; }
+void* DynamicLib_Get2(void* lib, const char* name) { return NULL; }
+cc_bool DynamicLib_DescribeError(cc_string* dst)   { return false; }
+
+
+/*########################################################################################################################*
+*--------------------------------------------------------Platform---------------------------------------------------------*
+*#########################################################################################################################*/
+EMSCRIPTEN_KEEPALIVE void Platform_LogError(const char* msg) {
+	/* no pointer showing more than 128 characters in chat */
+	cc_string str = String_FromRaw(msg, 128);
+	Logger_WarnFunc(&str);
+}
+
+extern void interop_InitModule(void);
+extern void interop_GetIndexedDBError(char* buffer);
+void Platform_Init(void) {
+	char tmp[64+1] = { 0 };
+	interop_InitModule();
+	
+	/* Check if an error occurred when pre-loading IndexedDB */
+	interop_GetIndexedDBError(tmp);
+	if (!tmp[0]) return;
+	
+	Chat_Add1("&cError preloading IndexedDB: %c", tmp);
+	Chat_AddRaw("&cPreviously saved settings/maps will be lost");
+	/* NOTE: You must pre-load IndexedDB before main() */
+	/* (because pre-loading only works asynchronously) */
+	/* If you don't, you'll get errors later trying to sync local to remote */
+	/* See doc/hosting-webclient.md for example preloading IndexedDB code */
+}
+
+
+/*########################################################################################################################*
+*-------------------------------------------------------Encryption--------------------------------------------------------*
+*#########################################################################################################################*/
+cc_result Platform_Encrypt(const void* data, int len, cc_string* dst) { return ERR_NOT_SUPPORTED; }
+cc_result Platform_Decrypt(const void* data, int len, cc_string* dst) { return ERR_NOT_SUPPORTED; }
+
+
+/*########################################################################################################################*
+*-----------------------------------------------------Configuration-------------------------------------------------------*
+*#########################################################################################################################*/
+int Platform_GetCommandLineArgs(int argc, STRING_REF char** argv, cc_string* args) {
+	int i, count;
+	argc--; argv++; /* skip executable path argument */
+
+	count = min(argc, GAME_MAX_CMDARGS);
+	for (i = 0; i < count; i++) { args[i] = String_FromReadonly(argv[i]); }
+	return count;
+}
+
+cc_result Platform_SetDefaultCurrentDirectory(int argc, char **argv) {
+	return chdir("/classicube") == -1 ? errno : 0;
+}
+#endif
